Added more analytics-like columns
- Moving this demo more towards being analytics-ey
- Note: we don't remove the old route column, because the data migration is asynchronous
  this resulted in errors like:

Cannot drop column route because it's affected by mutation with ID 'mutation_4.txt' which is not finished yet. Wait this mutation, or KILL it with command "KILL MUTATION WHERE mutation_id = 'mutation_4.txt'".

This is a critical area to investigate - what are the possibilities here?
Can we do migrations separately and wait for mutations?